+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - disable internal throttle
 - refactor OAuth token caching logic
 - code cleanup
+- improve error handling and logging
 
 ## [0.7.0] - 2025-10-28